+++ assets/example_file_with_stdio.h
@@ -1,0 +1,15 @@
+/* =========================================================================
+    Ceedling - Test-Centered Build System for C
+    ThrowTheSwitch.org
+    Copyright (c) 2010-26 Mike Karlesky, Mark VanderVoord, & Greg Williams
+    SPDX-License-Identifier: MIT
+========================================================================= */
+
+#ifndef EXAMPLE_FILE_WITH_STDIO_H
+#define EXAMPLE_FILE_WITH_STDIO_H
+
+#include <stdio.h>
+
+void print_number(FILE *f, int n);
+
+#endif /* EXAMPLE_FILE_WITH_STDIO_H */