+++ spec/system/gcov_deployment_spec.rb
@@ -33,6 +33,8 @@ ceedling_system_tests do
 
       test_case :project_with_gcov_success
       test_case :project_with_gcov_fail
+      test_case :gcov_console_report_with_system_header
+      test_case :gcov_console_report_with_partial
       # TODO: Restore these tests when the :abort_on_uncovered option is restored in the Gcov plugin
       # test_case :project_with_gcov_fail_because_of_uncovered_files
       # test_case :project_with_gcov_success_because_of_ignore_uncovered_list